day-4/ATS-Screening project/requirements.txt
streamlit
langgraph
langchain
langchain-openai
chromadb
pypdf
python-docx
python-dotenv
pandas
openai
